\name{dataset}
\Rdversion{1.1}
\alias{dataset}
\docType{data}
\title{
dataset simulated using the procedure described in the reference
}
\description{
First column is covariates, second outcomes, third and forth intrumental variables
}
\usage{data(dataset)}
\format{
  The format is:
 num [1:1000, 1:4] 0 1 1 0 0 1 1 0 0 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:4] "W" "Y" "" ""
}
\references{
A Study of Methods for Computing Empirical Likelihood
}
\examples{
data(dataset)
}
\keyword{datasets}
