\name{samp}
\Rdversion{1.1}
\alias{samp}
\docType{data}
\title{
sample from bootstrap
}
\description{
sample that is used in order to get the plot in the references
}
\usage{data(samp)}
\format{
  The format is:
 num [1:1000] 765 588 840 695 522 961 796 546 547 721 ...
}
\references{
A Study of Methods for Computing Empirical Likelihood
}
\examples{
data(samp)
}
\keyword{datasets}
