\name{vec}
\Rdversion{1.1}
\alias{vec}
\docType{data}
\title{
four constants
}
\description{
four constants that are used in order to get the plot in the reference
}
\usage{data(vec)}
\format{
  The format is:
 Named num [1:4] 11.8699 11.8957 4.184 -0.0348
 - attr(*, "names")= chr [1:4] "a1.temp" "a2.temp" "b1.temp" "d"
}
\references{
A Study of Methods for Computing Empirical Likelihood
}
\examples{
data(vec)
}
\keyword{datasets}
